+++ source/extensions/omni.isaac.franka/omni/isaac/franka/controllers/pick_place_controller.py
@@ -7,6 +7,7 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 import omni.isaac.motion_generation as mg
+from omni.isaac.core.articulations import Articulation
 from omni.isaac.franka.controllers import GripperController, RMPFlowController
 from typing import Optional, List
 
@@ -17,19 +18,25 @@ class PickPlaceController(mg.PickPlaceController):
         Args:
             name (str): [description]
             gripper_dof_indices (List[int]): [description]
-            robot_prim_path (str): [description]
+            robot_articulation (Articulation): [description]
             events_dt (Optional[List[float]], optional): [description]. Defaults to None.
         """
 
     def __init__(
-        self, name: str, gripper_dof_indices: List[int], robot_prim_path: str, events_dt: Optional[List[float]] = None
+        self,
+        name: str,
+        gripper_dof_indices: List[int],
+        robot_articulation: Articulation,
+        events_dt: Optional[List[float]] = None,
     ) -> None:
         if events_dt is None:
             events_dt = [0.008, 0.005, 1, 0.1, 0.05, 0.05, 0.0025, 1, 0.008, 0.08]
         mg.PickPlaceController.__init__(
             self,
             name=name,
-            cspace_controller=RMPFlowController(name=name + "_cspace_controller", robot_prim_path=robot_prim_path),
+            cspace_controller=RMPFlowController(
+                name=name + "_cspace_controller", robot_articulation=robot_articulation
+            ),
             gripper_controller=GripperController(
                 name=name + "_gripper_controller", gripper_dof_indices=gripper_dof_indices, deltas=None
             ),