+++ source/extensions/omni.isaac.motion_generation/omni/isaac/motion_generation/tests/__init__.py
@@ -7,4 +7,4 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 
-from .test_motion_generation import *
+from .test_motion_policy import *